Update to use Android Support 2.2.1
This commit ports Android Support SDK v2 to the Unity plugin. However, the iOS integration hasn't been touched - it's still using v1.11 of the Support SDK

This removes a lot of the responsibility of the old build/integration system in favour of using Unity's Gradle build. Instead of building the plugin source and copying across the AAR along with all of its AAR/JAR dependencies, we can publish the plugin artifact to a maven repo (in this case it uses a local tmp dir), and customise the Unity `mainTemplate.gradle` file to add that artifact as a dependency. This removes the need to copy across any Android artifacts directly. We also don't need to handle the duplicate CSS file issue anymore.

For theming, I have redeclared the Android Activity classes in the AndroidManifest and applied the `UnityTheme`. Hopefully this reduces theming issues for our integrators, as they are no longer required to set an AppCompat theme on their whole app.

This is a relatively simple integration with v2, focusing on compatibility with the Activities we provide in v2. The provider methods have not been tested, though they are largely the same as in v1.11.